requirements.txt
yfinance==0.2.36
pandas==2.0.1
pyyaml==6.0
aiohttp==3.9.1
prometheus_client==0.17.1
structlog==24.1.0
python-dotenv==1.0.0